discogs_credits: extract Discogs API helpers
Issue #32 — split out Discogs-specific code to src/api-discogs.js:

  link_infos                 the parsed-URL cache (still exported because
                             many consumers in the entry read it directly;
                             folds into a parser-returns-object shape in
                             a later chunk per ANALYSIS §2.3)
  getDiscogsLinkKey          the URL parser that populates link_infos
  _releaseDataCache          session-memory cache for release JSON
  getDiscogsReleaseData      fetches a release via api.discogs.com

~50 lines moved out. No behavior change; full 12-fixture suite green.
