+++ userscripts/discogs_credits/dev/DECISIONS.md
@@ -77,4 +77,5 @@ For the full reasoning behind any decision, `git log -- dev/ANALYSIS.md` finds t
 - 2026-05-24 21:50 — issue #32 chunk 1 (infrastructure) → `build.mjs` switched from pass-through copy to **esbuild bundle**. Metadata block extracted into `src/meta.txt` and prepended verbatim to the bundle. `src/discogs_credits.user.js` remains the lone entry module today (so this commit is no-behavior-change); subsequent chunks of issue #32 add new `src/*.js` modules and esbuild absorbs them automatically via the import graph. Bundle output drops from ~220 KB to ~155 KB (esbuild strips comments + collapses whitespace; src/ keeps full comments for ESLint and human reading). Test gate green on smoke fixture; full suite running. Watch mode now uses esbuild's `context().watch()` (tracks the module graph) with a side `fsWatch` on `meta.txt` (esbuild doesn't see non-JS inputs). Dev server's metadata-rewrite still works — it splits the served file at the `// ==/UserScript==` line and rewrites just the meta half.
 - 2026-05-24 22:05 — issue #32 chunk 2a (data tables) → extracted `ENTITY_TYPE_MAP`, `INSTRUMENTS`, and `WORK_ONLY_ARTIST_RELS` to `src/data/{entity-map,instruments,work-only-rels}.js` (~1180 lines total moved out of the 4983-line monolith). Imported from `src/discogs_credits.user.js`. `eslint.config.mjs` flipped `sourceType: 'script'` → `'module'` to accept the import statements. No behavior change. Smoke fixture passes; full suite running. These were the easiest moves: zero logic, just `export const`.
 - 2026-05-24 22:12 — issue #32 chunk 2b (constants + util) → extracted `REL_TEMPLATE`, `SELECTORS`, `DISCOGS_LOGO_URL` to `src/constants.js`; the five zero-dep DOM helpers (`doNext`, `setNativeValue`, `selectValue`, `makeKeyDownEvent`, `makeClickEvent`) to `src/util.js`. ~100 lines moved out. Entry file now down to 3713 lines (from 4983 originally; the original headroom was data + UI sugar). Smoke fixture passes; full suite running.
+- 2026-05-24 22:17 — issue #32 chunk 2c (Discogs API + URL helpers) → extracted `link_infos`, `getDiscogsLinkKey`, `_releaseDataCache`, and `getDiscogsReleaseData` to `src/api-discogs.js`. `link_infos` stays exported because many consumers in the entry still read `link_infos[key]` directly (this will move to a parser-returns-object shape per ANALYSIS §2.3 once the consumers themselves are moved). Smoke fixture passes; full suite running.
 - 2026-05-24 19:55 — issue #30 (entity-fetch slower than the old "5-thread burst" approach) → `mbThrottle` rewritten. The previous shape was a strict serial promise chain with a post-success `setTimeout(_gap)` between each request (`_gap` ∈ [200, 3000] ms), so the 5-worker pools in `checkMissingArtists` / `checkMissingCompanies` all funnelled through one queue at ≤5 req/s — slower than the old "5 immediate parallel threads" approach the throttle was meant to replace. New shape: up to `MAX_CONCURRENT=4` in-flight requests, no artificial gap, **global cooperative `_pauseUntil`** on 429/503 that honors `Retry-After` (or exponential per-request backoff if absent). All in-flight workers idle together until the pause elapses, which avoids the thundering-herd retries that the original 5-thread approach produced. Also removed two redundant `setTimeout(200)` calls (showReviewTable name pre-load loop and the URL-check pool) — the throttle paces. Expected ~3–5× speedup on rate-limited-heavy fixtures (Midwest Funk-style preflight), without 503 storms.
